requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==36.0.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
ecdsa==0.17.0
gunicorn==20.1.0
idna==3.3
numpy==1.21.5
Pillow==8.4.0
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.2.0
pymediainfo==5.1.0
python-jose==3.3.0
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
